[Refactor] MarketDetailView API 코드 수정 및 MarketCouponListView 수정
